finish the shimming, pull shim code into separate file
